autofetch the user previously logged in
